+++ volume_io/Volumes/input_nifti.c
@@ -289,7 +289,7 @@ initialize_nifti_format_input(VIO_STR             filename,
 {
   VIO_Status        status;
   int               sizes[VIO_MAX_DIMENSIONS];
-  int               n_voxels_in_slice;
+  long              n_voxels_in_slice;
   nc_type           desired_nc_type;
   int               axis;
   nifti_image       *nii_ptr;
@@ -420,7 +420,7 @@ initialize_nifti_format_input(VIO_STR             filename,
     if (axis < 3)
     {
       /* DEBUG */
-      printf("%d %d size:%4d step:%6.3f start:%9.4f dc:[%7.4f %7.4f %7.4f]\n", 
+      printf("%d %d size:%4d step:%6.3f start:%9.4f dc:[%7.4f %7.4f %7.4f]\n",
              axis,
              volume_axis,
              sizes[volume_axis],
@@ -506,8 +506,8 @@ input_more_nifti_format_file(
 
   if ( in_ptr->slice_index < in_ptr->sizes_in_file[2] )
   {
-    int        n_bytes_per_slice;
-    int        n_bytes_read;
+    long       n_bytes_per_slice;
+    size_t     n_bytes_read;
 
     n_bytes_per_slice = (in_ptr->sizes_in_file[0] *
                          in_ptr->sizes_in_file[1] *
